+++ kenallclient/models/v20230901.py
@@ -1,7 +1,7 @@
 """Models for API version 2023-09-01"""
 
 import dataclasses
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Optional
 
 __all__ = [
     "Corporation",
@@ -10,9 +10,6 @@ __all__ = [
     "AddressSearcherResponse",
     "City",
     "CityResolverResponse",
-    "NTACorporateInfo",
-    "NTACorporateInfoResolverResponse",
-    "NTACorporateInfoSearcherResponse",
     "Bank",
     "BankBranch",
     "BanksResponse",
@@ -161,127 +158,6 @@ class CityResolverResponse:
     def fromdict(cls, d: Dict[str, Any]) -> "CityResolverResponse":
         data = [City.fromdict(i) for i in d["data"]]
         return cls(version=d["version"], data=data)
-
-
-@dataclasses.dataclass()
-class NTACorporateInfo:
-    """Corporate info model for v2022-11-01"""
-
-    sequence_number: int
-    corporate_number: str
-    process: int
-    correct: int
-    update_date: str
-    change_date: str
-    name: str
-    name_image_id: Optional[str]
-    kind: int
-    prefecture_name: str
-    city_name: str
-    published_date: str
-    hihyoji: int
-    furigana: str
-    en_address_outside: Optional[str]
-    en_address_line: Optional[str]
-    en_prefecture_name: str
-    en_name: str
-    assignment_date: str
-    change_cause: str
-    successor_corporate_number: Optional[str]
-    close_cause: Optional[str]
-    close_date: Optional[str]
-    address_outside_image_id: Optional[str]
-    address_outside: str
-    post_code: str
-    jisx0402: str
-    address_image_id: Optional[str]
-    street_number: str
-    town: Optional[str]
-    kyoto_street: Optional[str]
-    block_lot_num: Optional[str]
-    building: Optional[str]
-    floor_room: Optional[str]
-
-    @classmethod
-    def fromdict(cls, d: Dict[str, Any]) -> "NTACorporateInfo":
-        return cls(**d)
-
-
-@dataclasses.dataclass()
-class NTACorporateInfoResolverResponse:
-    """Corporate info resolver response for v2022-11-01"""
-
-    version: str
-    data: NTACorporateInfo
-
-    @classmethod
-    def fromdict(cls, d: Dict[str, Any]) -> "NTACorporateInfoResolverResponse":
-        return cls(version=d["version"], data=NTACorporateInfo.fromdict(d["data"]))
-
-
-@dataclasses.dataclass()
-class NTACorporateInfoFacetResults:
-    area: Optional[List[Tuple[str, int]]]
-    kind: Optional[List[Tuple[str, int]]]
-    process: Optional[List[Tuple[str, int]]]
-    close_cause: Optional[List[Tuple[str, int]]]
-
-    def __getitem__(self, v: Any) -> List[Tuple[str, int]]:
-        if v == "area":
-            if self.area is not None:
-                return self.area
-        elif v == "kind":
-            if self.kind is not None:
-                return self.kind
-        elif v == "process":
-            if self.process is not None:
-                return self.process
-        elif v == "close_cause":
-            if self.close_cause is not None:
-                return self.close_cause
-        raise KeyError(v)
-
-    def __contains__(self, v: Any) -> bool:
-        if v == "area":
-            return self.area is not None
-        elif v == "kind":
-            return self.kind is not None
-        elif v == "process":
-            return self.process is not None
-        elif v == "close_cause":
-            return self.close_cause is not None
-        return False
-
-    @classmethod
-    def fromdict(cls, d: Dict[str, Any]) -> "NTACorporateInfoFacetResults":
-        return cls(
-            area=[tuple(pair) for pair in d["area"]] if "area" in d else None,
-            kind=[tuple(pair) for pair in d["kind"]] if "kind" in d else None,
-            process=[tuple(pair) for pair in d["process"]] if "process" in d else None,
-            close_cause=[tuple(pair) for pair in d["close_cause"]]
-            if "close_cause" in d
-            else None,
-        )
-
-
-@dataclasses.dataclass()
-class NTACorporateInfoSearcherResponse:
-    """Corporate info searcher response for v2022-11-01"""
-
-    version: str
-    data: List[str]
-    query: str
-    count: int
-    offset: int
-    limit: int
-    facets: List[List[Any]]
-
-    @classmethod
-    def fromdict(cls, d: Dict[str, Any]) -> "NTACorporateInfoSearcherResponse":
-        dd = dict(d)
-        dd["data"] = [NTACorporateInfo.fromdict(i) for i in dd["data"]]
-        dd["facets"] = NTACorporateInfoFacetResults.fromdict(dd.get("facets", {}))
-        return cls(**dd)
 
 
 # New models added in v2023-09-01 for Bank API
