+++ kenallclient/types.py
@@ -2,6 +2,4 @@
 
 from typing import Literal
 
-APIVersion = Literal[
-    "2022-09-01", "2022-11-01", "2023-09-01", "2024-01-01", "2025-01-01"
-]
+APIVersion = Literal["2022-11-01", "2023-09-01", "2024-01-01", "2025-01-01"]
