--- tests/fixtures/2022-09-01/city_get.json
@@ -1,14 +1,0 @@
-{
-  "version": "2022-09-01",
-  "data": [
-    {
-      "jisx0402": "13101",
-      "prefecture_code": "13",
-      "prefecture": "東京都",
-      "prefecture_kana": "トウキョウト",
-      "city_code": "101",
-      "city": "千代田区",
-      "city_kana": "チヨダク"
-    }
-  ]
-}
