--- tests/fixtures/2022-09-01/houjinbangou.json
@@ -1,39 +1,0 @@
-{
-  "version": "2022-09-01",
-  "data": {
-    "sequence_number": "1",
-    "corporate_number": "1234567890123",
-    "process": "1",
-    "correct": "0",
-    "update_date": "2023-01-01",
-    "change_date": "2023-01-01",
-    "name": "テスト株式会社",
-    "name_image_id": null,
-    "kind": "301",
-    "published_date": "2023-01-01",
-    "hihyoji": "0",
-    "furigana": "テストカブシキガイシャ",
-    "en_address_outside": null,
-    "en_address_line": "1-1-1 Otemachi, Chiyoda-ku",
-    "en_name": "Test Corporation",
-    "assignment_date": "2023-01-01",
-    "change_cause": "新規",
-    "successor_corporate_number": null,
-    "close_cause": "01",
-    "close_date": "2023-12-31",
-    "address_outside_image_id": null,
-    "address_outside": "",
-    "address_image_id": null,
-    "post_code": "1000001",
-    "jisx0402": "13101",
-    "prefecture_name": "東京都",
-    "en_prefecture_name": "Tokyo",
-    "city_name": "千代田区",
-    "street_number": "大手町１丁目１番１号",
-    "town": "大手町",
-    "kyoto_street": null,
-    "block_lot_num": "1-1-1",
-    "building": null,
-    "floor_room": null
-  }
-}
