--- tests/fixtures/2022-09-01/houjinbangou_search.json
@@ -1,134 +1,0 @@
-{
-  "version": "2022-09-01",
-  "data": [
-    {
-      "sequence_number": "1",
-      "corporate_number": "1234567890123",
-      "process": "1",
-      "correct": "0",
-      "update_date": "2022-01-01",
-      "change_date": "2022-01-01",
-      "name": "テスト株式会社",
-      "name_image_id": null,
-      "kind": "301",
-      "prefecture_name": "東京都",
-      "city_name": "港区",
-      "published_date": "2022-01-01",
-      "hihyoji": "0",
-      "furigana": "テストカブシキガイシャ",
-      "en_address_outside": null,
-      "en_address_line": "1-1-1 Akasaka, Minato-ku",
-      "en_prefecture_name": "Tokyo",
-      "en_name": "Test Corporation",
-      "assignment_date": "2022-01-01",
-      "change_cause": "",
-      "successor_corporate_number": null,
-      "close_cause": null,
-      "close_date": null,
-      "address_outside_image_id": null,
-      "address_outside": "",
-      "post_code": "1070052",
-      "jisx0402": "13103",
-      "address_image_id": null,
-      "street_number": "赤坂１丁目１番１号",
-      "town": "赤坂",
-      "kyoto_street": null,
-      "block_lot_num": "1-1-1",
-      "building": null,
-      "floor_room": null
-    },
-    {
-      "sequence_number": "2",
-      "corporate_number": "2345678901234",
-      "process": "1",
-      "correct": "0",
-      "update_date": "2022-02-01",
-      "change_date": "2022-02-01",
-      "name": "サンプル株式会社",
-      "name_image_id": null,
-      "kind": "301",
-      "prefecture_name": "東京都",
-      "city_name": "渋谷区",
-      "published_date": "2022-02-01",
-      "hihyoji": "0",
-      "furigana": "サンプルカブシキガイシャ",
-      "en_address_outside": null,
-      "en_address_line": "2-2-2 Shibuya, Shibuya-ku",
-      "en_prefecture_name": "Tokyo",
-      "en_name": "Sample Corporation",
-      "assignment_date": "2022-02-01",
-      "change_cause": "",
-      "successor_corporate_number": null,
-      "close_cause": null,
-      "close_date": null,
-      "address_outside_image_id": null,
-      "address_outside": "",
-      "post_code": "1500002",
-      "jisx0402": "13113",
-      "address_image_id": null,
-      "street_number": "渋谷２丁目２番２号",
-      "town": "渋谷",
-      "kyoto_street": null,
-      "block_lot_num": "2-2-2",
-      "building": null,
-      "floor_room": null
-    },
-    {
-      "sequence_number": "3",
-      "corporate_number": "3456789012345",
-      "process": "12",
-      "correct": "0",
-      "update_date": "2022-03-01",
-      "change_date": "2022-03-01",
-      "name": "大阪有限会社",
-      "name_image_id": null,
-      "kind": "302",
-      "prefecture_name": "大阪府",
-      "city_name": "大阪市北区",
-      "published_date": "2022-03-01",
-      "hihyoji": "0",
-      "furigana": "オオサカユウゲンガイシャ",
-      "en_address_outside": null,
-      "en_address_line": "3-3-3 Umeda, Kita-ku, Osaka-shi",
-      "en_prefecture_name": "Osaka",
-      "en_name": "Osaka Limited Company",
-      "assignment_date": "2021-01-01",
-      "change_cause": "国内所在地の変更",
-      "successor_corporate_number": null,
-      "close_cause": "11",
-      "close_date": "2022-12-31",
-      "address_outside_image_id": null,
-      "address_outside": "",
-      "post_code": "5300001",
-      "jisx0402": "27127",
-      "address_image_id": null,
-      "street_number": "梅田３丁目３番３号",
-      "town": "梅田",
-      "kyoto_street": null,
-      "block_lot_num": "3-3-3",
-      "building": null,
-      "floor_room": null
-    }
-  ],
-  "query": "name:テスト",
-  "count": 3,
-  "offset": 0,
-  "limit": 100,
-  "facets": {
-    "area": [
-      ["/東京都", 2],
-      ["/大阪府", 1]
-    ],
-    "kind": [
-      ["/株式会社", 2],
-      ["/有限会社", 1]
-    ],
-    "process": [
-      ["/新規", 2],
-      ["/国内所在地の変更", 1]
-    ],
-    "close_cause": [
-      ["/合併による解散等", 1]
-    ]
-  }
-}
