--- tests/fixtures/2022-09-01/postalcode_search.json
@@ -1,56 +1,0 @@
-{
-  "version": "2022-09-01",
-  "data": [
-    {
-      "jisx0402": "13101",
-      "old_code": "100",
-      "postal_code": "1000001",
-      "prefecture_kana": "トウキョウト",
-      "city_kana": "チヨダク",
-      "town_kana": "チヨダ",
-      "town_kana_raw": "チヨダ",
-      "prefecture": "東京都",
-      "city": "千代田区",
-      "town": "千代田",
-      "koaza": "",
-      "kyoto_street": "",
-      "building": "",
-      "floor": "",
-      "town_partial": false,
-      "town_addressed_koaza": false,
-      "town_chome": false,
-      "town_multi": false,
-      "town_raw": "千代田",
-      "corporation": null
-    },
-    {
-      "jisx0402": "13101",
-      "old_code": "102",
-      "postal_code": "1020072",
-      "prefecture_kana": "トウキョウト",
-      "city_kana": "チヨダク",
-      "town_kana": "イイダバシ",
-      "town_kana_raw": "イイダバシ",
-      "prefecture": "東京都",
-      "city": "千代田区",
-      "town": "飯田橋",
-      "koaza": "",
-      "kyoto_street": "",
-      "building": "",
-      "floor": "",
-      "town_partial": false,
-      "town_addressed_koaza": false,
-      "town_chome": true,
-      "town_multi": false,
-      "town_raw": "飯田橋",
-      "corporation": null
-    }
-  ],
-  "query": "千代田",
-  "count": 2,
-  "offset": 0,
-  "limit": 10,
-  "facets": {
-    "area": ["/東京都/千代田区", 2]
-  }
-}
